Initial commit, here's the source for all to see!
Here's the current version of WillyBot, which has been written on and off over the last two or three years, with a complete rewrite in the last month or so. There may be some residual odd code bits since this project was initially started in 2018, so blame past me (and current me). Hopefully releasing this much of the source will allow for those in my community to implement features which I am far too dumb to make.